Add the function of sending an order to VirtusizeRepository.kt
